Crate systemd-dbus

Rust API for Systemd commands via DBus.

See: http://www.freedesktop.org/wiki/Software/systemd/dbus/ http://dbus.freedesktop.org/doc/dbus-specification.html

Modules

serialize
Serialization support for data over DBus.

Structs

Connection
Main type representing a connection to systemd via Dbus.

Job
Systemd Job information

UnitAux
Unit properties

UnitFile
Systemd unit file information

UnitFileChange
Systemd unit file change information

UnitFileChanges
Systemd unit file changes

UnitProperty
A Unit property

UnitStatus
Unit status returned from systemd.

Enums

Match
Match rules for notifications

SystemdError
Errors that can arise on systemd DBus operations.

Functions

rule_string
Return a rule string for a sequence of matches

Type Definitions

ObjectPath
Define an ObjectPath as a String

SystemdResult
Result type for systemd DBus errors.