Module systemd-dbus::serialize

Serialization support for data over DBus.

Serialize between Rustc Encodable and Decodable types, and Vecdbus::MessageItem.

Structs

Decoder
Decoder

Encoder
Encoder object, to encode to Vec

Enums

DecoderError
Error type for serialization

EncoderError
Encoder Error Type

Functions

decode
Convenience function to decode a vector of MessageItem

encode
Utility function to encode an Encodable instance to a MessageItem

Type Definitions

DecodeResult
Result type for decoding dbus messages

EncodeResult
Result type for encoder functions.